comments fixup
